+++ docs/examples/controlpanels/delete_content_rule.ts
@@ -3,10 +3,7 @@ import { Client } from '@robgietema/nick';
 const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
 const login = await cli.login({ username: 'admin', password: 'admin' });
 
-const { data } = await cli.createControlpanelType({
+const { data } = await cli.deleteControlpanelContentRule({
   token: login.data.token,
-  data: {
-    title: 'My Type',
-    description: 'Type Description',
-  },
+  path: '/@controlpanels/content-rules/content-rule-1',
 });