+++ docs/examples/controlpanels/get_content_rules.ts
@@ -3,10 +3,4 @@ import { Client } from '@robgietema/nick';
 const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
 const login = await cli.login({ username: 'admin', password: 'admin' });
 
-const { data } = await cli.createControlpanelType({
-  token: login.data.token,
-  data: {
-    title: 'My Type',
-    description: 'Type Description',
-  },
-});
+const { data } = await cli.getControlpanelContentRules();